data_collection/requirements.txt
selenium==4.28.1
webdriver-manager==4.0.2
requests==2.28.1
Pillow==9.3.0
opencv-python==4.6.0.66
